requirements.txt
azure-functions
azure-batch==6.0.0
azure-storage-blob==1.4.0 
config